Cleanup Unix porting issues
git-svn-id: file:///home/jj/hercules.svn/trunk@6907 956126f8-22a0-4046-8f4a-272fa8102e63

+++ shared.c
@@ -2629,7 +2629,7 @@ va_list         vl;
     gettimeofday(&tv, NULL);
     sprintf ((char *)s,
             "%6.6ld" "." "%6.6ld %4.4X:",
-            tv.tv_sec, tv.tv_usec, dev ? dev->devnum : 0);
+            (long)tv.tv_sec, (long)tv.tv_usec, dev ? dev->devnum : 0);
     vsnprintf ((char *)s + strlen(s), sizeof(s) - strlen(s),
             msg, vl);
     if (dt)
